refactor: move utilities to utils directory
License: MIT
