+++ src/auto-launch.js
@@ -3,7 +3,7 @@ const os = require('os')
 const path = require('path')
 const fs = require('fs-extra')
 const untildify = require('untildify')
-const createToggler = require('./create-toggler')
+const createToggler = require('./utils/create-toggler')
 const logger = require('./common/logger')
 const store = require('./common/store')
 const { IS_MAC, IS_WIN } = require('./common/consts')
