--- src/dev-app-update.yml
@@ -1,3 +1,0 @@
-owner: ipfs-shipyard
-repo: ipfs-desktop
-provider: github
