+++ src/dialogs/prompt/template.js
