+++ src/ipfs-on-path/index.js
@@ -2,8 +2,8 @@ const { join } = require('path')
 const i18n = require('i18next')
 const which = require('which')
 const { execFile } = require('child_process')
-const createToggler = require('../create-toggler')
-const execOrSudo = require('../exec-or-sudo')
+const createToggler = require('../utils/create-toggler')
+const execOrSudo = require('../utils/exec-or-sudo')
 const logger = require('../common/logger')
 const store = require('../common/store')
 const { IS_WIN } = require('../common/consts')
