+++ test/unit/create-toggler.spec.js
@@ -22,7 +22,7 @@ describe('Create toggler', () => {
     store = mockStore()
     webui = mockWebUI()
     logger = mockLogger()
-    createToggler = proxyquire('../../src/create-toggler', {
+    createToggler = proxyquire('../../src/utils/create-toggler', {
       electron: electron,
       './common/logger': logger,
       './common/store': store
